Added backup restore and entry_move_to_another_category

+++ GUI/menu_functions/menu_functions.py
@@ -1,16 +1,32 @@
 from __future__ import annotations
+
+import os
 import time
-from pathlib import Path
+import threading
+import pyperclip
+import platform
+import subprocess
+from pathlib import Path 
+
+import logging
 
 from GUI.menu_functions.select_color_theme import SelectColorThemeFrame
-from GUI.modals.popups import dialog_popup, get_input, message_popup, save_file_as
+from GUI.modals.popups import dialog_popup, get_input, message_popup, save_file_as, select_dir, select_file
 from GUI.base_panel import BasePanel
 from GUI.modals.set_new_password import SetNewPasswordFrame
+from GUI.modals.copy_popup import CopyPopup
+from GUI.modals.get_password_from_user import GetPasswordFrame
 
 from typing import Optional, TYPE_CHECKING
 
 if TYPE_CHECKING:
     from GUI.base_panel import BasePanel
+    
+
+lock = threading.Lock()  
+logging.basicConfig(filename='app.log', level=logging.DEBUG)
+
+
 
 class MenuFunctions():
     
@@ -93,7 +109,7 @@ class MenuFunctions():
     def _name_is_valid(self, new_category):
         return not (new_category in self._base_panel._manage_data.data)
     
-    def add_category(self):
+    def add_category(self) -> None:
         color = self._base_panel._color_themes[self._base_panel._current_theme]['medium']
         new_category = get_input(color=color, hint="Enter desired category name:", title="New Category")
         if new_category is None or new_category == "":
@@ -140,13 +156,13 @@ class MenuFunctions():
     
     # Entry ------------------------------------------------------
     
-    def add_entry(self):
+    def add_entry(self) -> None:
         if self.manage_data.search_results is None:
             self.manage_data.add_entry()
             self._base_panel.refresh_mid_panel()
             self._base_panel.refresh_right_panel()
         
-    def remove_entry(self):
+    def remove_entry(self) -> None:
         if self.manage_data.selected_entry is None:
             return
         message = self._config['entry']['remove_dialog']['message']
@@ -163,35 +179,46 @@ class MenuFunctions():
             self._base_panel.refresh_mid_panel()
             self._base_panel.refresh_right_panel()
     
+    def move_entry_to_category(self, category: str):
+        if self.manage_data.move_entry_to_category(category=category):
+            BasePanel.set_selected_entry(None)
+            self._base_panel.refresh_body_panel()
+            message_popup(f"Entry moved to {category} successfully.", "Info.")
+    
     
     # Entry ------------------------------------------------------
     
     # State ------------------------------------------------------
     
-    def undo(self):
+    def undo(self) -> None:
         if self.manage_data.backward():
             self._base_panel.refresh_body_panel()
             self._base_panel.body.right_panel.deselect_all()
     
-    def redu(self):
+    def redu(self) -> None:
         if self.manage_data.forward():
             self._base_panel.refresh_body_panel() 
             self._base_panel.body.right_panel.deselect_all()
     
     # State ------------------------------------------------------
     
-    # State ------------------------------------------------------
+    # Search ------------------------------------------------------
     
-    def search(self):
+    def search(self) -> None:
         query = self._base_panel.top_panel.top_mid_panel.get_query()
-        self.manage_data.search(query)
         self.manage_data.search(query)
         self.manage_data.selected_category = None 
         self.manage_data.selected_entry = None 
         BasePanel.set_selected_entry(None)
         self._base_panel.refresh_body_panel()
-        
-    def choose_color_theme(self):
+
+    
+    # Search ------------------------------------------------------
+    
+    # State ------------------------------------------------------
+    
+
+    def choose_color_theme(self) -> None:
         if not self._color_panel_active:
             self._color_panel_active = True
             self.frame = SelectColorThemeFrame(self._base_panel, self.settings, self._base_panel._color_themes)
@@ -203,7 +230,7 @@ class MenuFunctions():
                 self._color_panel_active = False 
                 self.choose_color_theme()
     
-    def change_password(self):
+    def change_password(self) -> None:
         set_new_password = SetNewPasswordFrame(self._base_panel, self.manage_data._df, self.settings, self._base_panel._color_themes, self._base_panel._current_theme, change_password=True)
         set_new_password.Show()
     
@@ -212,13 +239,88 @@ class MenuFunctions():
     
     # Datafile ------------------------------------------------------
     
-    def save_datafile_as(self):
-        save_as = save_file_as("/", self._config['default_datafile_name'])
+    def save_datafile_as(self) -> None:
+        save_as = save_file_as(self._config['default_datafile_name'])
         if save_as is not None:
+            save_as = Path(save_as)
             datafile_path: Path = self.manage_data._df.io.file_path
-            with open(datafile_path, "r", encoding="utf-8") as f:
-                data = f.read()
-            with open(save_as, "w", encoding="utf-8") as f:
-                f.write(data)
-            message_popup(self._config['save_message']['message'].format(save_as), self._config['save_message']['title'])
+            if save_as.exists():
+                confirmed = dialog_popup(f"Aru you sure you want to replace it?", "File Already Exist")
+                if confirmed:
+                    with open(datafile_path, "r", encoding="utf-8") as f:
+                        data = f.read()
+                    with open(save_as, "w", encoding="utf-8") as f:
+                        f.write(data)
+                    message_popup(self._config['save_message']['message'].format(save_as), self._config['save_message']['title'])
+            else:
+                with open(datafile_path, "r", encoding="utf-8") as f:
+                        data = f.read()
+                with open(save_as, "w", encoding="utf-8") as f:
+                    f.write(data)
+                message_popup(self._config['save_message']['message'].format(save_as), self._config['save_message']['title'])
+            
+    def show_datafile_in_folder(self) -> None:
+        datafile_path: Path = self.manage_data._df.io.file_path
+        
+        if platform.system() == 'Windows':
+            subprocess.run(['explorer', '/select,', datafile_path])
+        elif platform.system() == 'Darwin':  # macOS
+            subprocess.run(['open', '-R', datafile_path])
+        else: 
+            subprocess.run(['xdg-open', os.path.dirname(datafile_path)])
+    
+    def copy_datafile_path(self) -> None:
+        datafile_path = str(self.manage_data._df.io.file_path.resolve())    
+        pyperclip.copy(datafile_path)   
+        copy_indicator = CopyPopup(self._base_panel) 
+        copy_indicator.Show()    
+    
+    
+    def change_datafile_dir(self) -> None:
+        datafile_path: Path = self.manage_data._df.io.file_path
+        file_name = datafile_path.name 
+        new_directory = select_dir()
+        if new_directory is not None:
+            new_path = Path(new_directory) / str(file_name)
+            if new_path.exists():
+                confirmed = dialog_popup(f"Aru you sure you want to replace it?", "File Already Exist")
+                if confirmed:
+                    self.manage_data._df.io.file_path = str(new_path) 
+                    self.manage_data.update()
+                return
+            
+            self.manage_data._df.io.file_path = str(new_path) 
+            self.manage_data.update()
+    
+    # def change_datafile(self) -> None:
+    #     self._current_file = str(self.manage_data._df.io.file_path)
+    #     self._current_password = self.manage_data._df.password 
+    #     new_datafile = select_file()
+    #     if new_datafile is not None:
+    #         self.manage_data._df.io.file_path = new_datafile
+    #         get_password = GetPasswordFrame(self.manage_data._df, self.settings, self._base_panel._color_themes, self._base_panel._current_theme, self)
+    #         get_password.Show()
+    
+    # def restore_df(self):
+    #     self.manage_data._df.io.file_path = self._current_file
+    #     self.manage_data._df.password = self._current_password
+    
+    # def LaunchMainApp(self) -> None:
+    #     self._base_panel.refresh_body_panel()
+    
+    def restore_from_backup(self) -> None:
+        current_file = str(self.manage_data._df.io.file_path)
+        backups_dir = self.settings['global']['backups']
+        restore_from = select_file(dir=backups_dir, title="Select Backup File")
+        if restore_from is not None:
+            self.manage_data._df.backup()
+            try:
+                self.manage_data._df.load_data(restore_from)
+            except ValueError:
+                message_popup(message=self.settings['get_password']['wrong_password']['message'], title=self.settings['get_password']['wrong_password']['title'])
+            else:
+                self._base_panel.refresh_body_panel()
+            
+            
+
